+++ tests/densities/test_fitting.py
@@ -126,11 +126,11 @@ def test_fill_in_homogenous_regions(hierarchy_info):
             "Basic cell groups and regions",
             "Central lobule",
             "Declive (VI)",
-            "Declive (VI), molecular layer",
-            "Declive (VI), Purkinje layer",
             "Lobule II",
             "Lobule II, granular layer",
             "Lobule II, Purkinje layer",
+            "Declive (VI), Purkinje layer",
+            "Declive (VI), molecular layer",
         ]
     )
     input_densities_dataframe = pd.DataFrame(
@@ -157,8 +157,8 @@ def test_fill_in_homogenous_regions(hierarchy_info):
     )
     expected = pd.DataFrame(
         {
-            "inhibitory_neuron": [np.nan, np.nan, 0.0, 0.0, 0.0, 0.4, 0.5, 0.4],
-            "inhibitory_neuron_standard_deviation": [np.nan, np.nan, 0.0, 0.0, 0.0, 0.4, 0.5, 0.4],
+            "inhibitory_neuron": [np.nan, np.nan, 0.0, 0.4, 0.5, 0.4, 0.0, 0.0],
+            "inhibitory_neuron_standard_deviation": [np.nan, np.nan, 0.0, 0.4, 0.5, 0.4, 0.0, 0.0],
             "pv+": np.full((8,), np.nan),
             "pv+_standard_deviation": np.full((8,), np.nan),
             "sst+": [0.1, np.nan, np.nan, np.nan, np.nan, np.nan, np.nan, np.nan],
@@ -204,7 +204,7 @@ def test_compute_average_intensity():
     volume_mask = np.zeros((2, 2, 2), dtype=bool)
     intensity = np.array([[[0.1, 0.1], [0.1, 0.2]], [[0.0, 0.1], [0.1, 0.1]]])
     actual = tested.compute_average_intensity(intensity, volume_mask)
-    assert np.isnan(actual)
+    assert actual == 0.0
 
     volume_mask = np.ones((2, 2, 2), dtype=bool)
     actual = tested.compute_average_intensity(intensity, volume_mask)
@@ -241,17 +241,17 @@ def test_compute_average_intensities(hierarchy_info):
             "Basic cell groups and regions",
             "Central lobule",
             "Declive (VI)",  # 936
-            "Declive (VI), molecular layer",
-            "Declive (VI), Purkinje layer",
             "Lobule II",  # 976
             "Lobule II, granular layer",
             "Lobule II, Purkinje layer",
+            "Declive (VI), Purkinje layer",
+            "Declive (VI), molecular layer",
         ]
     )
     expected = pd.DataFrame(
         {
-            "gad67": [7.0 / 4.0, 1.0, 2.0, np.nan, np.nan, 1.0, np.nan, np.nan],
-            "pv": [10.0 / 7.0, 2.0, 1.0, np.nan, np.nan, 2.0, np.nan, np.nan],
+            "gad67": [7.0 / 4.0, 1.0, 2.0, 1.0, 0.0, 0.0, 0.0, 0.0],
+            "pv": [10.0 / 7.0, 2.0, 1.0, 2.0, 0.0, 0.0, 0.0, 0.0],
         },
         index=hierarchy_info["brain_region"],
     )
@@ -278,17 +278,17 @@ def get_fitting_input_data_():
     h = get_hierarchy_info_()
     intensities = pd.DataFrame(
         {
-            "gad67": [7.0 / 4.0, 1.0, 2.0, np.nan, np.nan, 1.0, np.nan, np.nan],
-            "pv": [10.0 / 7.0, 2.0, 1.0, np.nan, np.nan, 2.0, np.nan, np.nan],
+            "gad67": [7.0 / 4.0, 1.0, 2.0, 1.0, np.nan, np.nan, np.nan, np.nan],
+            "pv": [10.0 / 7.0, 2.0, 1.0, 2.0, np.nan, np.nan, np.nan, np.nan],
         },
         index=h["brain_region"],
     )
     densities = pd.DataFrame(
         {
-            "gad67+": [7.0 / 2.0, 2.0, 4.0, np.nan, np.nan, 2.0, np.nan, np.nan],
-            "gad67+_standard_deviation": [1.0, 0.0, 3.0, np.nan, np.nan, 0.0, np.nan, np.nan],
-            "pv+": [30.0 / 7.0, 6.0, 3.0, np.nan, np.nan, 6.0, np.nan, np.nan],
-            "pv+_standard_deviation": [2.0, 1.0, 3.0, np.nan, np.nan, 4.0, np.nan, np.nan],
+            "gad67+": [7.0 / 2.0, 2.0, 4.0, 2.0, np.nan, np.nan, np.nan, np.nan],
+            "gad67+_standard_deviation": [1.0, 0.0, 3.0, 0.0, np.nan, np.nan, np.nan, np.nan],
+            "pv+": [30.0 / 7.0, 6.0, 3.0, 6.0, np.nan, np.nan, np.nan, np.nan],
+            "pv+_standard_deviation": [2.0, 1.0, 3.0, 4.0, np.nan, np.nan, np.nan, np.nan],
         },
         index=h["brain_region"],
     )
@@ -361,27 +361,27 @@ def test_fit_unknown_densities(hierarchy_info, fitting_coefficients):
     }
     intensities = pd.DataFrame(
         {
-            "gad67": [7.0 / 4.0, 1.0, 2.0, np.nan, 1.0, 1.0, 1.5, np.nan],
-            "pv": [10.0 / 7.0, 2.0, 1.0, np.nan, 0.5, 2.0, 0.75, np.nan],
+            "gad67": [7.0 / 4.0, 1.0, 2.0, 1.0, 1.5, np.nan, 1.0, np.nan],
+            "pv": [10.0 / 7.0, 2.0, 1.0, 2.0, 0.75, np.nan, 0.5, np.nan],
         },
         index=hierarchy_info["brain_region"],
     )
     densities = pd.DataFrame(
         {
-            "gad67+": [7.0 / 2.0, 2.0, 4.0, np.nan, np.nan, 2.0, np.nan, np.nan],
-            "gad67+_standard_deviation": [1.0, 0.0, 3.0, np.nan, np.nan, 0.0, np.nan, np.nan],
-            "pv+": [30.0 / 7.0, 6.0, 3.0, np.nan, np.nan, 6.0, np.nan, np.nan],
-            "pv+_standard_deviation": [2.0, 1.0, 3.0, np.nan, np.nan, 4.0, np.nan, np.nan],
+            "gad67+": [7.0 / 2.0, 2.0, 4.0, 2.0, np.nan, np.nan, np.nan, np.nan],
+            "gad67+_standard_deviation": [1.0, 0.0, 3.0, 0.0, np.nan, np.nan, np.nan, np.nan],
+            "pv+": [30.0 / 7.0, 6.0, 3.0, 6.0, np.nan, np.nan, np.nan, np.nan],
+            "pv+_standard_deviation": [2.0, 1.0, 3.0, 4.0, np.nan, np.nan, np.nan, np.nan],
         },
         index=hierarchy_info["brain_region"],
     )
 
     expected = pd.DataFrame(
         {
-            "gad67+": [7.0 / 2.0, 2.0, 4.0, np.nan, 2.0, 2.0, 3.0, np.nan],
-            "gad67+_standard_deviation": [1.0, 0.0, 3.0, np.nan, 2.0, 0.0, 1.5, np.nan],
-            "pv+": [30.0 / 7.0, 6.0, 3.0, np.nan, 1.5, 6.0, 2.25, np.nan],
-            "pv+_standard_deviation": [2.0, 1.0, 3.0, np.nan, 0.75, 4.0, 2.25, np.nan],
+            "gad67+": [7.0 / 2.0, 2.0, 4.0, 2.0, 3.0, np.nan, 2.0, np.nan],
+            "gad67+_standard_deviation": [1.0, 0.0, 3.0, 0.0, 1.5, np.nan, 2.0, np.nan],
+            "pv+": [30.0 / 7.0, 6.0, 3.0, 6.0, 2.25, np.nan, 1.5, np.nan],
+            "pv+_standard_deviation": [2.0, 1.0, 3.0, 4.0, 2.25, np.nan, 0.75, np.nan],
         },
         index=hierarchy_info["brain_region"],
     )